+++ titiler/openeo/processes/implementations/apply.py
@@ -1,6 +1,6 @@
 """titiler.openeo.processes Apply."""
 
-from typing import Callable, Dict, Optional, Union
+from typing import Callable, Dict, Optional
 
 from .data_model import ImageData, RasterStack
 
@@ -8,11 +8,11 @@ __all__ = ["apply"]
 
 
 def apply(
-    data: Union[RasterStack, ImageData],
+    data: RasterStack,
     process: Callable,
     context: Optional[Dict] = None,
-) -> Union[RasterStack, ImageData]:
-    """Apply process on Data."""
+) -> RasterStack:
+    """Apply process on RasterStack."""
     positional_parameters = {"x": 0}
     named_parameters = {"context": context}
 
@@ -28,7 +28,5 @@ def apply(
             bounds=img.bounds,
         )
 
-    if isinstance(data, ImageData):
-        return _process_img(data)
-
+    # Apply process to each item in the stack
     return {k: _process_img(img) for k, img in data.items()}